Allow Failovers to be scheduled in the future
Manual failovers may be scheduled using the api. The 'planned_at' value of the failover will
be interpreted as a timestamp. To avoid any ambiguity in the timestamp, the timezone will be enriched
with timezone information if it lacks this information. Patronictl will take the timezone from the client,
the api will take it from the running Patroni.

If a failover key contains the 'planned_at' value, the timestamp will be evaluated. If it is deemed to be
in the future, the function processing the failover returns immediately. If planned roughly equates to
now, the failover commences. Planned failovers in the past are considered stale and will be removed when found.