Use ssl as a default for replication if enabled
Some setups require all network connections to use SSL.
As the first pg_hba entry is allowing non-ssl replication connections,
we could not enforce this requirement.

This patch changes the behaviour of Patroni for clusters running SSL.
If replication connections without ssl enabled are still to be allowed
to the cluster, they should be explicitly specified in the Patroni configuration.

Some tests were added to ensure we don't write hostssl records if ssl is disabled.