Refactor demotion logic and remove wait_for_failover.
On further reflection about the behavior of failover after master recovery timeout, we actually don't need or want to wait for failover before initiating startup as replica. The 2 second grace period used in manual failover is enough. So merge all demote() cases into a single function with a mode switch.

There is a race condition here in that we will forget to rewind, but that is shared with manual failover and is a separate issue. The race condition is that somebody that is behind the current node promotes during the 2 second grace period. The race condition is rather more likely here because:

* the window when we are not eligible for promotion is larger due to recovery process.
* we didn't wait for connected standbys to receive all WAL.

The fix may be better fixed by not resetting the failover flag when starting up as a standby.