Lift setting of need_rewind out of follow to simplify the API a bit.
Small change of behavior - in paused mode if we ever discover that we are in an unhealthy cluster without a failover key pointing to us, we used to do a rewind on the next iteration if there was a leader then, never otherwise. After this change we do a rewind only when there is a leader available, don't clear the flag before that.