Add postgres-9.6 support
starting from 9.6 we need wal_level = 'replica' which is alias for
'hot_standby'. It was working before without problems, but if somebody
change wal_level to replica, Patroni will expose pending_restart flag,
alghough restart in this case is not necessary.

Besides that bump versions of consul and etcd to the latest.