Export locale variables when forking a postmaster.
Avoid "postmaster became multithreaded during startup" error on OS X
built with --enable-nls (default for petere/homebrew).

The issue is that on OS X the libintl that replaces setlocale() spawns
a thread when it needs to detect the current locale. Setting the LC_
variables prevents this, but we don't propagate them to the fork.

See:
https://www.postgresql.org/message-id/flat/20140902013458.GB906981%40tornado.leadboat.com#20140902013458.GB906981@tornado.leadboat.com
for an additional explanation of the original Postgres behavior.