Remove background keepalives, simplify watchdog logic
In the only two cases that were left using background keepalives we want
to disable watchdog anyway as soon as it safe to do so. This allows for
the removal of the complicated BackgroundKeepaliveSender, replacing the
background thread and signaling logic with a simple callback.

In addition, because now there is only one place left sending keepalives
we don't have any chance of duplicated keepalives so we can remove the
deduplication logic.

Needed to add synchronization on the watchdog, because a failover task
could call it from the async thread. Trivial method level
synchronization should be fine as we don't expect any contention here.